+++ core-ng/src/main/java/core/framework/api/http/HTTPHeaders.java
@@ -7,4 +7,5 @@ public final class HTTPHeaders {
     public static final String ACCEPT = "Accept";
     public static final String CONTENT_TYPE = "Content-Type";
     public static final String LOCATION = "Location";
+    public static final String IF_MODIFIED_SINCE = "If-Modified-Since";
 }