+++ core-ng/src/main/java/core/framework/api/module/SiteConfig.java
@@ -1,11 +1,16 @@
 package core.framework.api.module;
 
 import core.framework.impl.module.ModuleContext;
+import core.framework.impl.web.StaticContentController;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * @author neo
  */
 public class SiteConfig {
+    private final Logger logger = LoggerFactory.getLogger(SiteConfig.class);
+
     private final ModuleContext context;
 
     public SiteConfig(ModuleContext context) {
@@ -18,5 +23,10 @@ public class SiteConfig {
 
     public void template(String path, Class<?> modelClass) {
         context.httpServer.templateManager.addTemplate(path, modelClass);
+    }
+
+    public void staticContent(String root) {
+        logger.info("add static content root, root={}", root);
+        context.httpServer.get(root + "/:path(*)", new StaticContentController(context.httpServer.webDirectory, root));
     }
 }