+++ core-ng/src/main/java/core/framework/impl/web/HTTPServer.java
@@ -30,8 +30,9 @@ public class HTTPServer implements RouteConfig, HTTPConfig {
 
     private final Logger logger = LoggerFactory.getLogger(HTTPServer.class);
 
+    public final WebDirectory webDirectory = new WebDirectory();
     public final SessionManager sessionManager = new SessionManager();
-    public final TemplateManager templateManager = new TemplateManager();
+    public final TemplateManager templateManager = new TemplateManager(webDirectory);
     public final BeanValidator validator = new BeanValidator();
     public final WebContextImpl webContext = new WebContextImpl();
     private final Route route = new Route();