+++ core-ng/src/main/java/core/framework/impl/web/TemplateManager.java
@@ -7,11 +7,6 @@ import core.framework.impl.template.TemplateBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Map;
 
 /**
@@ -20,31 +15,10 @@ import java.util.Map;
 public class TemplateManager {
     private final Logger logger = LoggerFactory.getLogger(TemplateManager.class);
     private final Map<String, Template> templates = Maps.newConcurrentHashMap();
-    private final Path webPath;
-    private boolean refreshTemplateIfModified;
+    private final WebDirectory webDirectory;
 
-    public TemplateManager() {
-        webPath = locateWebPath();
-    }
-
-    private Path locateWebPath() {
-        String webPathValue = System.getProperty("core.web");
-        if (webPathValue != null) {
-            Path webPath = Paths.get(webPathValue).toAbsolutePath();
-            if (Files.exists(webPath) && Files.isDirectory(webPath)) {
-                logger.info("found -Dcore.web, use it as web path, path={}", webPath);
-                return webPath;
-            }
-        } else {
-            Path webPath = Paths.get("./src/main/dist/web").toAbsolutePath();
-            if (Files.exists(webPath) && Files.isDirectory(webPath)) {
-                logger.warn("found local web path, this should only happen in local dev env, path={}", webPath);
-                refreshTemplateIfModified = true;
-                return webPath;
-            }
-        }
-        logger.info("can not locate web path");
-        return null;
+    public TemplateManager(WebDirectory webDirectory) {
+        this.webDirectory = webDirectory;
     }
 
     public String process(String templateName, Object model) {
@@ -68,22 +42,14 @@ public class TemplateManager {
     }
 
     private Template loadTemplate(String templateName, Class<?> modelClass) {
-        Path templatePath = webPath.resolve(templateName).toAbsolutePath();
-        logger.debug("load template, path={}", templatePath);
-        String template = core.framework.api.util.Files.text(templatePath);
+        logger.debug("load template, path={}", templateName);
+        String template = webDirectory.text(templateName);
         return new TemplateBuilder(template, modelClass).build();
     }
 
     private String templateKey(String templateName) {
-        if (webPath == null)
-            throw new Error("web path does not exist, check -Dcore.web or set working dir to be module path for local dev env.");
-
-        if (refreshTemplateIfModified) {
-            try {
-                return templateName + ":" + Files.getLastModifiedTime(webPath.resolve(templateName)).toMillis();
-            } catch (IOException e) {
-                throw new UncheckedIOException(e);
-            }
+        if (webDirectory.localEnv) {
+            return templateName + ":" + webDirectory.lastModified(templateName).toMillis();
         } else {
             return templateName;
         }