+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.2.7 (7/21/2015)
 * db repository, added selectAll()
+* template engine first draft impl, removed thymeleaf
 
 ### 3.2.6 (7/17/2015 - 7/20/2015)
 * refactory rabbitmq support