+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Change log
 
-### 3.2.7 (7/21/2015)
+### 3.2.7 (7/21/2015 - 7/22/2015)
 * db repository, added selectAll()
 * template engine first draft impl, removed thymeleaf
+* website static content
 
 ### 3.2.6 (7/17/2015 - 7/20/2015)
 * refactory rabbitmq support